+++ Engine7/Engine/Sandbox/Logic/World/Tiles/CompositeTile.cs
@@ -8,18 +8,18 @@ public sealed class CompositeTile( Icosphere icosphere, CompositeTile? containin
 	public int IndexA { get; } = indexA;
 	public int IndexB { get; } = indexB;
 	public int IndexC { get; } = indexC;
-	public Vector3<double> VectorA => _icosphere.Vertices[ IndexA ];
-	public Vector3<double> VectorB => _icosphere.Vertices[ IndexB ];
-	public Vector3<double> VectorC => _icosphere.Vertices[ IndexC ];
+	public Vector3<float> VectorA => _icosphere.Vertices[ IndexA ];
+	public Vector3<float> VectorB => _icosphere.Vertices[ IndexB ];
+	public Vector3<float> VectorC => _icosphere.Vertices[ IndexC ];
 	public uint RemainingLayers { get; private set; } = 0;
 	public uint Layer { get; } = layer;
 	public ITile? ContainingTile { get; } = containingTile;
 	public IReadOnlyList<ITile> SubTiles { get; private set; } = [];
 
-	public Vector4<double> Color => GetColor();
+	public Vector4<float> Color => GetColor();
 
-	private Vector4<double> GetColor() {
-		Vector4<double> color = Vector4<double>.AdditiveIdentity;
+	private Vector4<float> GetColor() {
+		Vector4<float> color = Vector4<float>.AdditiveIdentity;
 		foreach (ITile subTile in SubTiles)
 			color += subTile.Color;
 		color /= SubTiles.Count;
